Day 9 ruby solutions.
Should have been around 5:30 or so for second place, but I thought there
were only six cities, not seven, and hardcoded 6 for that.  After a bit, I
checked on the example and changed it to actually use $dists.size, and then
it mysteriously worked.  It wasn't until after getting part 2 as well that I found the actual error.  *sigh*.